pyproject.toml
[project]
name = "multi-website-scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "hf-xet>=1.4.2",
    "ipywidgets>=8.1.8",
    "langchain>=1.2.12",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-docling>=2.0.0",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "lxml>=6.0.2",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
]
